+++ stelvio/aws/appsync/appsync.py
@@ -622,6 +622,7 @@ class AppSync(Component[AppSyncResources, AppSyncCustomizationDict], LinkableMix
             domain_name=self._config.domain,
             tags=self.tags,
             customize=self._customize.get("acm_validated_domain"),
+            parent=self,
         )
 
         domain_name = appsync.DomainName(