+++ stelvio/aws/cloudfront/router.py
@@ -75,6 +75,7 @@ class Router(Component[RouterResources, RouterCustomizationDict]):
                 tags=self.tags,
                 customize=self._customize.get("acm_validated_domain"),
                 region="us-east-1",
+                parent=self,
             )
 
         if not self.routes: