+++ tests/aws/test_customization_sync.py
@@ -157,7 +157,7 @@ from tests.test_utils import assert_resources_matches_customization_dict
         pytest.param(
             RestApiResources,
             RestApiCustomizationDict,
-            {"dns_record"},
+            {"dns_record", "acm_validated_domain"},
             None,
             id="RestApi",
         ),